+++ src/app/core/@ngrx/root-store.module.ts
@@ -1,0 +1,33 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { StoreModule } from '@ngrx/store';
+import { ProductsStoreModule } from './products/products-store.module';
+import { EffectsModule } from '@ngrx/effects';
+import { RouterEffects } from './router/router.effects';
+import { RouterState, StoreRouterConnectingModule } from '@ngrx/router-store';
+import { routerReducers } from './router/router.reducers';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    StoreModule.forRoot(routerReducers, {
+      // All checks will automatically be disabled in production builds
+      runtimeChecks: {
+        strictStateImmutability: true,
+        strictActionImmutability: true,
+        strictStateSerializability: false,
+        strictActionSerializability: false,
+        strictActionWithinNgZone: true,
+        strictActionTypeUniqueness: true
+      }
+    }),
+    StoreRouterConnectingModule.forRoot({
+      stateKey: 'router',
+      routerState: RouterState.Minimal
+    }),
+    EffectsModule.forRoot([RouterEffects]),
+    ProductsStoreModule,
+  ]
+})
+export class RootStoreModule { }