+++ src/app/core/@ngrx/router/index.ts
@@ -1,0 +1,5 @@
+export * from './router.actions';
+export * from './router.effects';
+export * from './router.reducers';
+export * from './router.selectors';
+export * from './router.state';