+++ src/app/core/@ngrx/shp.state.ts
@@ -1,0 +1,5 @@
+import { ProductsState } from './products';
+
+export interface ShpState {
+  products: ProductsState;
+}