+++ src/app/features/products/components/product/product.component.ts
@@ -1,6 +1,9 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { Store } from '@ngrx/store';
+import { ShpState } from 'src/app/core/@ngrx';
 import { CartService } from 'src/app/features/cart/services/cart.service';
 import { ProductModel } from '../../models/product';
+import * as RouterActions from './../../../../core/@ngrx/router/router.actions';
 
 @Component({
   selector: 'shp-product',
@@ -8,6 +11,10 @@ import { ProductModel } from '../../models/product';
   styleUrls: ['./product.component.scss']
 })
 export class ProductComponent {
+
+  constructor(
+    private store: Store<ShpState>,
+  ) { }
 
   @Input()
   product: ProductModel;
@@ -17,5 +24,12 @@ export class ProductComponent {
 
   onBuyProduct(): void {
     this.buyProduct.emit(this.product.id);
+  }
+
+  onViewProduct(id: number): void {
+    const link = ['/product', id];
+    this.store.dispatch(RouterActions.go({
+      path: link
+    }));
   }
 }