+++ src/app/shared/pipes/order-by.pipe.ts
@@ -6,12 +6,13 @@ import { Pipe, PipeTransform } from '@angular/core';
 })
 export class OrderByPipe implements PipeTransform {
 
-  transform<T>(array: T[] | null, keys: Array<keyof T>, isAsc: boolean = true): T[] | null {
+  transform<T>(array: T[] | ReadonlyArray<T> | null, keys: Array<keyof T>, isAsc: boolean = true): T[] | ReadonlyArray<T> | null {
     if (!Array.isArray(array)) {
       return array;
     }
+    const copyArray = [...array]; // required for Readonly arrays
     const orderingValue = isAsc ? 1 : -1;
-    return array.sort((a, b) => {
+    return copyArray.sort((a, b) => {
       for (const key of keys) {
         if (a[key] > b[key])
           return orderingValue;